refactor(anim): drop doc refs from comments; mod.rs export-only
Code-style cleanup:
- Comments no longer point at documentation. Removed every `docs/...` path, plan
  section label (`§x.y`), invariant tag (`I1`–`I6`), and milestone/decision
  citation from comments across the editor + renderer-animation code, rewriting
  each to stand on its own (stating the rule in plain words; referencing code
  symbols where useful). Also dropped the now-meaningless milestone tag from a
  user-facing import toast.
- mod.rs is export-only: moved the Animation-mode workspace layout out of
  animation_mode/mod.rs into workspace.rs, and the shared timeline geometry +
  track-target labelling out of timeline/mod.rs into timeline/shared.rs. Both
  mod.rs now just declare modules + re-export.
- Deleted the two completed plan docs (animation-editor.md, editor-rewrite.md);
  the forward-looking scene-loader.md stays.

No behavior change. task lint green; 39 animation unit tests pass.
